Add C api version for native components
 #### Summary of Changes
For native components read the C api version information from
the bundle similarly to what we do for "platform" information.
The C api version is specific to each bundle.

If the bundle doesn't contain a C api version, then return "N/A"
and skip this information in the manifest file. Otherwise use
the C api version to populate the manifest inside appPackageManifest.

 #### Description of testing
Checked the changes are backwards compatible by generating a manifest
for a native component without C api version.
Checked that the manifest contains the C api version information
if it is present in the native component.